+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.0.0
+
+Bumped this straight to version 1.0.0
+
+Created a listener for the`addStickerPack` method
+Added a new build config field for the sticker_packs.json file
+Added propper documentation to the plugin dart file
+
+Updated the plugin import statements in the example app
+Added a title to example app
+
+Updated the readme
+Updated .gitignore
+
 ## 0.0.1
 
 * Initial Release